Improve introduction
Yield away from referring to Pluto as a 'fork', as you'd expect forks to only add small things and by now we're quite past that.
Forget about explaining what Lua is, we need to pack a punch in this because this is what google's crawlers put in the result embeds.